pyproject.toml
[tool.poetry]
name = "llama-synth"
version = "0.1.0"
description = ""
authors = ["hjlee94 <infin94@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
llama_cpp_python = "^0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "run:classify_label"
